website info/backend/requirements.txt
# FastAPI and core dependencies
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.3
pydantic-settings==2.2.1
python-multipart==0.0.9
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
email-validator==2.1.0.post1

# Database
sqlalchemy==2.0.27
alembic==1.13.1
aiosqlite==0.19.0
psycopg2-binary==2.9.9  # PostgreSQL adapter

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# API Documentation
# fastapi-swagger-ui==2.0.0  # Not available - using FastAPI's built-in swagger

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.23.3
pytest-mock==3.12.0

# Utilities
requests==2.31.0
python-slugify==8.0.1
Pillow==10.2.0
tenacity==8.2.3
pytz==2024.1

# Payment Processing
stripe==7.9.0
squareup  # Updated square package name
paypalrestsdk==1.13.1
qrcode==7.4.2

# PDF Generation
reportlab==4.0.8
xhtml2pdf==0.2.15

# Email services
# Uncommented SendGrid for email integration
sendgrid==6.11.0

# Email
jinja2==3.1.3  # For email templates

# Environment & Configuration
python-dotenv==1.0.1

# Optional: Redis for session storage (if needed later)
# redis==5.0.1

# Optional: Email service (if using SendGrid, Mailgun, etc.)
# mailgun==0.1.1

# Supabase Integration
pyjwt[crypto]==2.8.0

# Dev Tools
black==23.12.1
isort==5.13.2
mypy==1.7.1 